requirements.txt
dify_plugin>=0.2.0,<0.3.0
requests>=2.31.0
PyJWT>=2.8.0
